增加了反馈和评价两部分功能

+++ src/main/java/com/example/feiyang/entity/Evaluation.java
@@ -4,11 +4,11 @@ import java.io.Serializable;
 import java.util.Date;
 
 public class Evaluation implements Serializable {
-    private Integer evaluationId;
+    private Long evaluationId;
 
-    private Integer orderId;
+    private Long orderId;
 
-    private Integer userId;
+    private Long userId;
 
     private Date createTime;
 
@@ -18,27 +18,27 @@ public class Evaluation implements Serializable {
 
     private static final long serialVersionUID = 1L;
 
-    public Integer getEvaluationId() {
+    public Long getEvaluationId() {
         return evaluationId;
     }
 
-    public void setEvaluationId(Integer evaluationId) {
+    public void setEvaluationId(Long evaluationId) {
         this.evaluationId = evaluationId;
     }
 
-    public Integer getOrderId() {
+    public Long getOrderId() {
         return orderId;
     }
 
-    public void setOrderId(Integer orderId) {
+    public void setOrderId(Long orderId) {
         this.orderId = orderId;
     }
 
-    public Integer getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(Integer userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
